# rest/src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: enviroCar REST API
  description: enviroCar REST API
  version: "1.0.0"
  contact:
    email: info@envirocar.org
    name: "52\xB0North GmbH"
    url: https://52north.org
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://envirocar.org/conditions.html?lng=en
servers: [ ]
...
paths:
...
  '/sensors':
    get:
      operationId: getSensors
      summary: Get the Sensor listing.
      description: Get the Sensor listing.
      tags:
        - Sensors
      parameters:
        - $ref: '#/components/parameters/sensorType'
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/range'
      responses:
        '200':
          $ref: '#/components/responses/Sensors'
        '206':
          $ref: '#/components/responses/Sensors'
        '400':
          $ref: '#/components/responses/BadRequest'
        '5XX':
          $ref: '#/components/responses/InternalServerError'
    post:
      operationId: createSensor
      summary: Create a new Sensor.
      description: Create a new Sensor.
      tags:
        - Sensors
      requestBody:
        $ref: '#/components/requestBodies/CreateSensor'
      responses:
        '201':
          $ref: '#/components/responses/Created'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '451':
          $ref: '#/components/responses/UnavailableForLegalReasons'
        '5XX':
          $ref: '#/components/responses/InternalServerError'
  '/sensors/{sensorId}':
    parameters:
      - $ref: '#/components/parameters/sensorId'
    get:
      operationId: getSensorById
      summary: Get the Sensor by id.
      description: Get the Sensor by id.
      tags:
        - Sensors
      responses:
        '200':
          $ref: '#/components/responses/Sensor'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '5XX':
          $ref: '#/components/responses/InternalServerError'
...
  '/tracks':
    get:
      operationId: getTracks
      summary: Get the list of tracks.
      description: Get the list of tracks.
      tags:
        - Tracks
      parameters:
        - $ref: '#/components/parameters/bbox'
        - $ref: '#/components/parameters/nearPoint'
        - $ref: '#/components/parameters/after'
        - $ref: '#/components/parameters/before'
        - $ref: '#/components/parameters/begins'
        - $ref: '#/components/parameters/begunBy'
        - $ref: '#/components/parameters/contains'
        - $ref: '#/components/parameters/during'
        - $ref: '#/components/parameters/endedBy'
        - $ref: '#/components/parameters/ends'
        - $ref: '#/components/parameters/equals'
        - $ref: '#/components/parameters/meets'
        - $ref: '#/components/parameters/metBy'
        - $ref: '#/components/parameters/overlapped'
        - $ref: '#/components/parameters/overlaps'
        - $ref: '#/components/parameters/status'
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/range'
      responses:
        '200':
          $ref: '#/components/responses/Tracks'
        '206':
          $ref: '#/components/responses/Tracks'
        '400':
          $ref: '#/components/responses/BadRequest'
        '5XX':
          $ref: '#/components/responses/InternalServerError'
    post:
      operationId: createTrack
      summary: Create a new track.
      description: Create a new track.
      tags:
        - Tracks
      security:
        - basicAuth: [ ]
      requestBody:
        $ref: '#/components/requestBodies/CreateTrack'
      responses:
        '201':
          $ref: '#/components/responses/Created'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/Forbidden'
        '451':
          $ref: '#/components/responses/UnavailableForLegalReasons'
        '5XX':
          $ref: '#/components/responses/InternalServerError'
  '/tracks/{trackId}':
    parameters:
      - $ref: '#/components/parameters/trackId'
    get:
      operationId: getTrackById
      summary: Get the specified track.
      description: Get the specified track.
      tags:
        - Tracks
      responses:
        '200':
          $ref: '#/components/responses/Track'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '5XX':
          $ref: '#/components/responses/InternalServerError'
    put:
      operationId: modifyTrack
      summary: Update the specified track.
      description: Update the specified track.
      tags:
        - Tracks
      security:
        - basicAuth: [ ]
      requestBody:
        $ref: '#/components/requestBodies/ModifyTrack'
      responses:
        '204':
          $ref: '#/components/responses/NoContent'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '5XX':
          $ref: '#/components/responses/InternalServerError'
    delete:
      operationId: deleteTrack
      summary: Delete the specified track.
      description: Delete the specified track.
      tags:
        - Tracks
      security:
        - basicAuth: [ ]
      responses:
        '204':
          $ref: '#/components/responses/NoContent'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '5XX':
          $ref: '#/components/responses/InternalServerError'
...
  '/measurements':
    get:
      operationId: getMeasurements
      summary: Get the list of measurements.
      description: Get the list of measurements.
      tags:
        - Measurements
      parameters:
        - $ref: '#/components/parameters/bbox'
        - $ref: '#/components/parameters/nearPoint'
        - $ref: '#/components/parameters/after'
        - $ref: '#/components/parameters/before'
        - $ref: '#/components/parameters/begins'
        - $ref: '#/components/parameters/begunBy'
        - $ref: '#/components/parameters/contains'
        - $ref: '#/components/parameters/during'
        - $ref: '#/components/parameters/endedBy'
        - $ref: '#/components/parameters/ends'
        - $ref: '#/components/parameters/equals'
        - $ref: '#/components/parameters/meets'
        - $ref: '#/components/parameters/metBy'
        - $ref: '#/components/parameters/overlapped'
        - $ref: '#/components/parameters/overlaps'
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/range'
      responses:
        '200':
          $ref: '#/components/responses/Measurements'
        '206':
          $ref: '#/components/responses/Measurements'
        '400':
          $ref: '#/components/responses/BadRequest'
        '5XX':
          $ref: '#/components/responses/InternalServerError'
    post:
      operationId: createMeasurement
      summary: Create a new measurement.
      description: Create a new measurement.
      tags:
        - Measurements
      security:
        - basicAuth: [ ]
      requestBody:
        $ref: '#/components/requestBodies/CreateMeasurement'
      responses:
        '201':
          $ref: '#/components/responses/Created'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/Forbidden'
        '451':
          $ref: '#/components/responses/UnavailableForLegalReasons'
        '5XX':
          $ref: '#/components/responses/InternalServerError'
...
components:
  parameters:
    after:
      description: The time instant or interval to be applied with the `after` operator.
      in: query
      name: after
      required: false
      schema:
        $ref: '#/components/schemas/TemporalFilter'
    bbox:
      description: The bounding box the requested tracks have to intersect with.
      explode: false
      in: query
      name: bbox
      required: false
      style: form
      schema:
        items:
          type: number
        maxItems: 4
        minItems: 4
        type: array
    before:
      description: The time instant or interval to be applied with the `before` operator.
      in: query
      name: before
      required: false
      schema:
        $ref: '#/components/schemas/TemporalFilter'
    begins:
      description: The time instant or interval to be applied with the `begins` operator.
      in: query
      name: begins
      required: false
      schema:
        $ref: '#/components/schemas/TemporalFilter'
    begunBy:
      description: The time instant or interval to be applied with the `begunBy` operator.
      in: query
      name: begunBy
      required: false
      schema:
        $ref: '#/components/schemas/TemporalFilter'
    contains:
      description: The time instant or interval to be applied with the `contains`
        operator.
      in: query
      name: contains
      required: false
      schema:
        $ref: '#/components/schemas/TemporalFilter'
    during:
      description: The time instant or interval to be applied with the `during` operator.
      in: query
      name: during
      required: false
      schema:
        $ref: '#/components/schemas/TemporalFilter'
    endedBy:
      description: The time instant or interval to be applied with the `endedBy` operator.
      in: query
      name: endedBy
      required: false
      schema:
        $ref: '#/components/schemas/TemporalFilter'
    ends:
      description: The time instant or interval to be applied with the `ends` operator.
      in: query
      name: ends
      required: false
      schema:
        $ref: '#/components/schemas/TemporalFilter'
    equals:
      description: The time instant or interval to be applied with the `equals` operator.
      in: query
      name: equals
      required: false
      schema:
        $ref: '#/components/schemas/TemporalFilter'
...
    meets:
      description: The time instant or interval to be applied with the `meets` operator.
      in: query
      name: meets
      required: false
      schema:
        $ref: '#/components/schemas/TemporalFilter'
    metBy:
      description: The time instant or interval to be applied with the `metBy` operator.
      in: query
      name: metBy
      required: false
      schema:
        $ref: '#/components/schemas/TemporalFilter'
    nearPoint:
      description: The point to which the distance can not fall below a certain threshold.
      explode: false
      in: query
      name: nearPoint
      required: false
      schema:
        items:
          type: number
        maxItems: 3
        minItems: 3
        type: array
      style: form
    overlapped:
      description: The time instant or interval to be applied with the `overlapped`
        operator.
      in: query
      name: overlapped
      required: false
      schema:
        $ref: '#/components/schemas/TemporalFilter'
    overlaps:
      description: The time instant or interval to be applied with the `overlaps`
        operator.
      in: query
      name: overlaps
      required: false
      schema:
        $ref: '#/components/schemas/TemporalFilter'
    status:
      in: query
      name: status
      required: false
      schema:
        $ref: 'schema/track-status.json'
    limit:
      description: The number of entities per page.
      in: query
      name: limit
      required: false
      schema:
        type: integer
        default: 100
    page:
      description: The page of entities to request.
      in: query
      name: page
      required: false
      schema:
        type: integer
        default: 1
...
    range:
      description: The range of entities to request.
      in: header
      name: Range
      required: false
      style: simple
      explode: true
      schema:
        type: object
        properties:
          items:
            type: string
            pattern: '^\d+-\d+$'
...
    sensorId:
      description: The Sensor id.
      in: path
      name: sensorId
      required: true
      schema:
        type: string
    sensorType:
      description: The Sensor type.
      in: query
      name: type
      required: false
      schema:
        type: string
    trackId:
      description: The track id.
      in: path
      name: trackId
      required: true
      schema:
        type: string
...
  headers:
    ContentRange:
      style: simple
      explode: true
      schema:
        type: string
        pattern: '^items\s+\d+-\d+/(\d+|\*)$'
    Link:
      style: simple
      explode: true
      schema:
        type: object
        properties:
          items:
            type: string
            pattern: '^<.*>\s?;\s?rel=(first|last|next|prev|service-desc)\s?;\s?type=.+/.+'
    Location:
      schema:
        type: string
        format: uri
    WWW-Authenticate:
      schema:
        type: string
  requestBodies:
    CreateMeasurement:
      description: The measurement to create.
      required: true
      content:
        application/json:
          schema:
            $ref: 'schema/measurement-create.json'
...
    CreateSensor:
      description: The sensor to create.
      required: true
      content:
        application/json:
          schema:
            $ref: 'schema/sensor-create.json'
    CreateTrack:
      description: The track to create.
      required: true
      content:
        application/json:
          schema:
            $ref: 'schema/track-create.json'
...
    ModifyTrack:
      description: The track modifications.
      required: true
      content:
        application/json:
          schema:
            $ref: 'schema/track-modify.json'
...
  responses:
    NoContent:
      description: A empty response
      headers:
        Link:
          $ref: '#/components/headers/Link'
    Created:
      description: The entity was created.
      headers:
        Location:
          $ref: '#/components/headers/Location'
        Link:
          $ref: '#/components/headers/Link'
...
    BadRequest:
      description: The request was malformed.
      content:
        application/json:
          schema:
            $ref: 'schema/exception.json'
      headers:
        Link:
          $ref: '#/components/headers/Link'
    Forbidden:
      description: The operation is not allowed by current user.
      content:
        application/json:
          schema:
            $ref: 'schema/exception.json'
      headers:
        Link:
          $ref: '#/components/headers/Link'
    InternalServerError:
      description: An internal server error occured.
      content:
        application/json:
          schema:
            $ref: 'schema/exception.json'
      headers:
        Link:
          $ref: '#/components/headers/Link'
    NotFound:
      description: The Resource could not be found.
      content:
        application/json:
          schema:
            $ref: 'schema/exception.json'
      headers:
        Link:
          $ref: '#/components/headers/Link'
    UnavailableForLegalReasons:
      description: The operation requieres acceptance of the terms of use and privacy statement.
      content:
        application/json:
          schema:
            $ref: 'schema/exception.json'
      headers:
        Link:
          $ref: '#/components/headers/Link'
    UnauthorizedError:
      description: Authentication information is missing or invalid
      content:
        application/json:
          schema:
            $ref: 'schema/exception.json'
      headers:
        Link:
          $ref: '#/components/headers/Link'
        WWW-Authenticate:
          $ref: '#/components/headers/WWW-Authenticate'
...
    Sensor:
      description: The Sensor.
      content:
        application/json:
          schema:
            $ref: 'schema/sensor.json'
        text/yaml:
          schema:
            $ref: 'schema/sensor.json'
      headers:
        Link:
          $ref: '#/components/headers/Link'
    Sensors:
      description: The Sensors.
      content:
        application/json:
          schema:
            $ref: 'schema/sensors.json'
        text/yaml:
          schema:
            $ref: 'schema/sensors.json'
      headers:
        Link:
          $ref: '#/components/headers/Link'
        Content-Range:
          $ref: '#/components/headers/ContentRange'
...
    Track:
      description: The track.
      content:
        application/json:
          schema:
            $ref: 'schema/track.json'
        text/yaml:
          schema:
            $ref: 'schema/track.json'
      headers:
        Link:
          $ref: '#/components/headers/Link'
    Tracks:
      description: The tracks.
      content:
        application/json:
          schema:
            $ref: 'schema/tracks.json'
        text/yaml:
          schema:
            $ref: 'schema/tracks.json'
      headers:
        Link:
          $ref: '#/components/headers/Link'
        Content-Range:
          $ref: '#/components/headers/ContentRange'
...
    Measurements:
      description: The measurements.
      content:
        application/json:
          schema:
            $ref: 'schema/measurements.json'
        text/yaml:
          schema:
            $ref: 'schema/measurements.json'
      headers:
        Link:
          $ref: '#/components/headers/Link'
        Content-Range:
          $ref: '#/components/headers/ContentRange'
...
  schemas:
    TemporalFilter:
      oneOf:
        - type: string
          description: Two comma-seperated date-times that form a temporal interval.
        - format: date-time
          type: string
          description: The temporal instant.
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic